+++ text_extensions_for_pandas/array/tensor.py
@@ -71,7 +71,7 @@ class TensorType(pd.api.extensions.ExtensionDtype):
         return TensorArray
 
     def __from_arrow__(self, extension_array):
-        from text_extensions_for_pandas.array.arrow_compat import ArrowTensorArray
+        from text_extensions_for_pandas.array.arrow_conversion import ArrowTensorArray
         values = ArrowTensorArray.to_numpy(extension_array)
         return TensorArray(values)
 
@@ -233,7 +233,7 @@ class TensorArray(pd.api.extensions.ExtensionArray, TensorOpsMixin):
             raise NotImplementedError(f"'{name}' aggregate not implemented.")
 
     def __arrow_array__(self, type=None):
-        from text_extensions_for_pandas.array.arrow_compat import ArrowTensorArray
+        from text_extensions_for_pandas.array.arrow_conversion import ArrowTensorArray
         return ArrowTensorArray.from_numpy(self._tensor)
